Change members directory nav-filter element id to be sure ajax filters the display
After a deeper look into code, it seems the ids expected need to be component-scope. For instance, the group nav filter are groups-all and groups-personal. After checking css, there is no rules for members-show-all or members-show-personal. To be consistent with groups compnent i think it is best to name these nav filters members-all and members-personal.

Fixes #76